Add a GranvilleRule class to analyze the signal, detail not implemented

+++ ma-crossover/ma.c
@@ -1,9 +1,9 @@
 typedef struct {
   long f1;
-  double f2;
-  double f3;
-  double f4;
-  double f5;
+  float f2;
+  float f3;
+  float f4;
+  float f5;
 } stRawData;
 
 __kernel void ma(__global unsigned int* raw,
@@ -22,11 +22,11 @@ __kernel void test_donothing(const int aRange,
   if (startIdx < 0) {
     startIdx = 0;
   }
-  unsigned int nCount = (gid-startIdx)+1;
-  double f2temp = 0.0;
-  double f3temp = 0.0;
-  double f4temp = 0.0;
-  double f5temp = 0.0;
+  float nCount = (gid-startIdx)+1;
+  float f2temp = 0.0;
+  float f3temp = 0.0;
+  float f4temp = 0.0;
+  float f5temp = 0.0;
   for (int i = startIdx; i <= gid; i++) {
     f2temp += aIn[i].f2;
     f3temp += aIn[i].f3;
